requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-admin==1.6.1
werkzeug==3.0.1
python-dotenv==1.0.0
flask-migrate==4.0.5
